lengthCnt must be 16 Bit, one full tone for 160/4 has 375

+++ sound-driver/sequencer.h
@@ -38,7 +38,7 @@ typedef enum {
 
 typedef struct {
   uint16_t idx;
-  uint8_t lengthCnt;
+  uint16_t lengthCnt;
   t_sequencerState state;
   uint8_t amplitude;
   const t_tone *tones;
